Fixed undefined index issues and combat allowances
Many indexes from the database weren't included in the spell editing. I
don't know what some of them do, so a new section in the bottom was
created called 'Values that need assignment'. Removed query for
$spellname, see comment. Commented out spellcats, see comment. Fixed
check for the checkboxes, if a checkbox isn't checked, it doesn't get
sent through $_POST, so have to use isset to not give an error.